Rename some file (#338)
* rename test

* rename file

* rename file

* fix import in test

* fix example
